# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DEPTH Procedural API",
    "version": "1.0.0",
    "description": "Local, stateless, deterministic generation for DEPTH 0.5.0. Node 20+. No key, external AI, persistence or hosted endpoint. Requests are capped at 64 KiB and batches at 20 rooms. Bound to 127.0.0.1; same-origin HTTP requests or local scripts only. HEAD also supported for catalog and health routes. No automatic physics reachability check."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api/v1"
    }
  ],
  "paths": {
...
    "/catalog": {
      "get": {
        "operationId": "getCatalog",
        "summary": "Complete game catalog",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or nonlocal Host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Unsupported method",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 65536 bytes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Expected application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/levels/generate": {
      "get": {
        "operationId": "generateLevelFromQuery",
        "summary": "Generate one reproducible room",
        "description": "Duplicate parameters are rejected. Query integer syntax is decimal digits only.",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "story",
                "expedition"
              ],
              "default": "expedition"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,48}$",
              "default": "LUMA",
              "description": "Expedition seed, normalized to uppercase. Ignored in story."
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100000,
              "default": 1,
              "description": "Story maximum 100. Cannot exceed a nonzero expedition length."
            }
          },
          {
            "name": "intensity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 5,
              "default": 1,
              "description": "Expedition only. Story normalizes to 1."
            }
          },
          {
            "name": "length",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                12,
                36
              ],
              "default": 12,
              "description": "0 means Abismo, capped at 100000 rooms. Story normalizes to 12."
            }
          },
          {
            "name": "worldId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "vestibule",
                "gardens",
                "archives",
                "nexus",
                "core"
              ],
              "description": "Omit to mix worlds. Available in expedition v5 only."
            }
          },
          {
            "name": "generatorVersion",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                4,
                5
              ],
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedLevel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or nonlocal Host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Unsupported method",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 65536 bytes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Expected application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "generateLevel",
        "summary": "Generate one reproducible room",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              },
              "example": {
                "mode": "expedition",
                "seed": "LUMA-RECUERDA",
                "level": 1,
                "intensity": 3,
                "length": 36,
                "worldId": "archives",
                "generatorVersion": 5
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedLevel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or nonlocal Host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Unsupported method",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 65536 bytes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Expected application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/levels/batch": {
      "post": {
        "operationId": "generateBatch",
        "summary": "Generate 1–20 consecutive rooms",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or nonlocal Host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Unsupported method",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 65536 bytes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Expected application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/levels/validate": {
      "post": {
        "operationId": "validateWorld",
        "summary": "Check structural bounds, geometry and puzzle references",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "world"
                ],
                "properties": {
                  "world": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Foreign origin or nonlocal Host",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown route",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "405": {
            "description": "Unsupported method",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds 65536 bytes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Expected application/json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GenerateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "story",
              "expedition"
            ],
            "default": "expedition"
          },
          "seed": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,48}$",
            "default": "LUMA",
            "description": "Expedition seed, normalized to uppercase. Ignored in story."
          },
          "level": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000,
            "default": 1,
            "description": "Story maximum 100. Cannot exceed a nonzero expedition length."
          },
          "intensity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "default": 1,
            "description": "Expedition only. Story normalizes to 1."
          },
          "length": {
            "type": "integer",
            "enum": [
              0,
              12,
              36
            ],
            "default": 12,
            "description": "0 means Abismo, capped at 100000 rooms. Story normalizes to 12."
          },
          "worldId": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              null,
              "vestibule",
              "gardens",
              "archives",
              "nexus",
              "core"
            ],
            "default": null,
            "description": "Explicit world requires expedition and generatorVersion 5. Null mixes all worlds."
          },
          "generatorVersion": {
            "type": "integer",
            "enum": [
              4,
              5
            ],
            "default": 5
          }
        },
        "examples": [
          {
            "mode": "expedition",
            "seed": "LUMA-RECUERDA",
            "level": 1,
            "intensity": 3,
            "length": 36,
            "worldId": "archives",
            "generatorVersion": 5
          }
        ],
        "allOf": [
          {
            "if": {
              "properties": {
                "mode": {
                  "const": "story"
                }
              },
              "required": [
                "mode"
              ]
            },
            "then": {
              "properties": {
                "level": {
                  "maximum": 100
                },
                "worldId": {
                  "type": "null"
                }
              }
            }
          },
          {
            "if": {
              "properties": {
                "generatorVersion": {
                  "const": 4
                }
              },
              "required": [
                "generatorVersion"
              ]
            },
            "then": {
              "properties": {
                "worldId": {
                  "type": "null"
                }
              }
            }
          }
        ]
      },
      "BatchRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "story",
              "expedition"
            ],
            "default": "expedition"
          },
          "seed": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,48}$",
            "default": "LUMA",
            "description": "Expedition seed, normalized to uppercase. Ignored in story."
          },
          "level": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100000,
            "default": 1,
            "description": "Story maximum 100. Cannot exceed a nonzero expedition length."
          },
          "intensity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "default": 1,
            "description": "Expedition only. Story normalizes to 1."
          },
          "length": {
            "type": "integer",
            "enum": [
              0,
              12,
              36
            ],
            "default": 12,
            "description": "0 means Abismo, capped at 100000 rooms. Story normalizes to 12."
          },
          "worldId": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              null,
              "vestibule",
              "gardens",
              "archives",
              "nexus",
              "core"
            ],
            "default": null,
            "description": "Explicit world requires expedition and generatorVersion 5. Null mixes all worlds."
          },
          "generatorVersion": {
            "type": "integer",
            "enum": [
              4,
              5
            ],
            "default": 5
          },
          "count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20,
            "default": 1
          }
        },
        "description": "Same rules as GenerateRequest. Generates count consecutive rooms beginning at level; rejects the entire request if any room exceeds the run length."
      },
      "Validation": {
        "type": "object",
        "required": [
          "valid",
          "errors",
          "reachability"
        ],
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reachability": {
            "type": "string",
            "const": "not-run"
          }
        },
        "description": "Structural checks only. No physics solver, safety proof or human difficulty rating."
      },
      "GeneratedLevel": {
        "type": "object",
        "required": [
          "schemaVersion",
          "generatorVersion",
          "id",
          "parameters",
          "world",
          "validation"
        ],
        "properties": {
          "schemaVersion": {
            "const": 1
          },
          "generatorVersion": {
            "enum": [
              4,
              5
            ]
          },
          "id": {
            "type": "string"
          },
          "parameters": {
            "$ref": "#/components/schemas/GenerateRequest"
          },
          "world": {
            "$ref": "#/components/schemas/World"
          },
          "validation": {
            "$ref": "#/components/schemas/Validation"
          }
        }
      },
      "Batch": {
        "type": "object",
        "required": [
          "schemaVersion",
          "count",
          "levels"
        ],
        "properties": {
          "schemaVersion": {
            "const": 1
          },
          "count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20
          },
          "levels": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "$ref": "#/components/schemas/GeneratedLevel"
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "status",
              "message"
            ],
            "properties": {
              "status": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Catalog": {
        "type": "object",
        "required": [
          "schemaVersion",
          "generatorVersions",
          "worlds",
          "chapters",
          "levels",
          "characters",
          "abilities",
          "obstacles",
          "rewards",
          "memories"
        ],
        "properties": {
          "schemaVersion": {
            "const": 1
          },
          "generatorVersions": {
            "type": "array",
            "items": {
              "enum": [
                4,
                5
              ]
            }
          },
          "worlds": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id"
              ]
            }
          },
          "chapters": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id"
              ]
            }
          },
          "levels": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id"
              ]
            }
          },
          "characters": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id"
              ]
            }
          },
          "abilities": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id"
              ]
            }
          },
          "obstacles": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id"
              ]
            }
          },
          "rewards": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id"
              ]
            }
          },
          "memories": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id"
              ]
            }
          }
        }
      }
    }
  }
}